Backport ruby-openssl 2.0.9 from Ruby 2.4.9 to Ruby 1.9-2.3
The openssl extension in Ruby 1.8-2.3 no longer works with
LibreSSL after the switch to opaque structures.  Backporting
the version from 2.4.9 allows the openssl to continue to work.
However, it requires extension patching to get it to compile.

This is barely tested beyond compiling and some basic net/https
testing.  However, it is probably better than having no openssl
extension at all.

For ruby 1.8, do not build the openssl extension.  I didn't look
into how much work it would be to backport it, but my guess is
too much.