Fix OpenSSL::HMAC.digest after LibreSSL change
Previous code called HMAC function with a NULL buffer, which caused
LibreSSL to return a pointer to a static buffer.  Support for this
was removed from LibreSSL in June 2024.  Allocate a stack buffer
and pass that instead of NULL.  Also, add error handling, which
was previously missing and would result in an empty string
being returned.